+++ leafset/leafset.go
@@ -1,0 +1,6 @@
+package leafset
+
+// LeafSet contains the sets of numerically closer and farther from the node.
+type LeafSet struct {
+	smaller, larger Set
+}
